refactor(Storage): revert abstract provider to use unsafe storage connection.
Reader API does not provide any safe indication to properly control position.

To work around the problem, we should create a sort of proxy to count invokations of mark/reset on the reader. For now, we'll keep using an unsafe approach for this type of objects, as they should not be used heavily.